requirements/base.txt
asgiref==3.7.2
Django==4.2.4
django-ckeditor==6.7.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-js-asset==2.1.0
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
